+++ Scripts/Persistence/Models/Market.cs
@@ -7,7 +7,7 @@ public class Market : _Entity{
     public long EntryDate{ get; set; }
     public string Code{ get; set; }
 
-    public long IdMarket{ get; set; }
+    public int IdMarket{ get; set; }
     public List<int> Items{ get; set; } // IDs de Items asociados al market
     public List<int> ItemsPurchase{ get; set; } // IDs de los items de compra (compra real)
 }
